finished dialog box. Just need to fix the percentage number.

+++ app/src/main/java/cs4330/cs/utep/edu/PriceFinder.java
@@ -5,9 +5,9 @@ import java.util.Random;
 public class PriceFinder {
 
     public float findPrice(){
-        double price = 0.00;
+        float price = (float) 0.00;
         Random r = new Random();
-        price = (r.nextInt((int)((500-50)*10+1))+50*10) / 10.0;
+        price = (float) ((r.nextInt((int)((500-50)*10+1))+50*10) / 10.0);
 
         float floatingnumber = (float) price;
         return floatingnumber;
